feat: rely on Playwright web server for tests

+++ packages/wb/src/scripts/execution/httpServerScripts.ts
@@ -35,22 +35,25 @@ class HttpServerScripts extends BaseScripts {
   }
 
   override startTest(project: Project, argv: ScriptArgv): string {
+    if (!this.shouldLaunchTestServer(project)) {
+      return this.skipServerStartCommand();
+    }
     return `${project.hasPrisma ? 'prisma migrate reset --force --skip-generate && ' : ''}(${this.startProduction(
       project,
       argv
     )})`;
   }
 
-  override testE2E(
-    project: Project,
-    argv: TestArgv,
-    {
+  override testE2E(project: Project, argv: TestArgv, options: TestE2EOptions = {}): string {
+    if (project.hasWebServerOnPlaywrightConfig()) {
+      return super.testE2E(project, argv, options);
+    }
+    const {
       startCommand = `${project.hasPrisma ? 'prisma migrate reset --force --skip-generate && ' : ''}(${this.startProduction(
         project,
         argv
       )})`,
-    }: TestE2EOptions
-  ): string {
+    } = options;
     const port = project.env.PORT || '8080';
     const suffix = project.packageJson.scripts?.['test/e2e-additional'] ? ' && YARN test/e2e-additional' : '';
     const testTarget = argv.targets && argv.targets.length > 0 ? argv.targets.join(' ') : 'test/e2e/';
@@ -59,7 +62,11 @@ class HttpServerScripts extends BaseScripts {
       "wait-on -t 600000 -i 2000 http-get://127.0.0.1:${port} && vitest run ${testTarget} --color --passWithNoTests --allowOnly${suffix}"`;
   }
 
-  override testE2EDev(project: Project, argv: TestArgv, { startCommand }: TestE2EDevOptions): string {
+  override testE2EDev(project: Project, argv: TestArgv, options: TestE2EDevOptions = {}): string {
+    if (project.hasWebServerOnPlaywrightConfig()) {
+      return super.testE2EDev(project, argv, options);
+    }
+    const { startCommand } = options;
     const port = project.env.PORT || '8080';
     const suffix = project.packageJson.scripts?.['test/e2e-additional'] ? ' && YARN test/e2e-additional' : '';
     const testTarget = argv.targets && argv.targets.length > 0 ? argv.targets.join(' ') : 'test/e2e/';
@@ -71,6 +78,9 @@ class HttpServerScripts extends BaseScripts {
   }
 
   override async testStart(project: Project, argv: ScriptArgv): Promise<string> {
+    if (!this.shouldLaunchTestServer(project)) {
+      return this.skipServerStartCommand();
+    }
     const port = await findAvailablePort();
     return `WB_ENV=${project.env.WB_ENV} PORT=${port} YARN concurrently --kill-others --raw --success first "${this.start(project, argv)}" "${this.waitApp(project, argv, port)}"`;
   }
